test: Separate tests into files and cond testing
