+++ docs/make.jl
@@ -5,9 +5,11 @@ using MakieMaestro
 
 # NOTE: When updating, must update also in `test/runtests.jl` and `test/fix_doctests.jl`<18-12-24> 
 DocMeta.setdocmeta!(
-    MakieMaestro, :DocTestSetup, :(using MakieMaestro;
-    using Logging; # NOTE: This does not need to be in the `make.jl` of docs. We want `@warn ` to function there <19-12-24> 
-    Logging.disable_logging(Logging.Warn)); recursive=true
+    MakieMaestro, :DocTestSetup, :(
+        include(joinpath(@__DIR__, "..", "test/doctestsetup.jl"));
+        using MakieMaestro;
+        using Logging; # NOTE: This does not need to be in the `make.jl` of docs. We want `@warn ` to function there <19-12-24> 
+        Logging.disable_logging(Logging.Warn)); recursive=true
 )
 
 # NOTE: Links can be explored trough the REPL with `links(query)`
