+++ test/doctestsetup.jl
@@ -1,0 +1,1 @@
+using MakieMaestro
